+++ CHANGELOG.md
@@ -18,11 +18,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 
+
+## [3.8.0] - NOT YET RELEASED
+[3.8.0]:https://github.com/secuconnect/secuconnect-java-sdk/compare/3.7.1...3.8.0
+
+### Added
+- `Api.PaymentTransactionsApi.getPaymentTransactionsCrowdFundingData()`
+- `Model.PaymentContext` element `creditcard_schemes`
+- `Model.SmartTransactionsApplicationContext` element `language`
+
+
 ## [3.7.1] - 2021-12-13
 [3.7.1]:https://github.com/secuconnect/secuconnect-java-sdk/compare/3.7.0...3.7.1
 
 ### Added
 - Using the Apache Maven registry from GitHub.
+
 
 ## [3.7.0] - 2021-12-13
 [3.7.0]:https://github.com/secuconnect/secuconnect-java-sdk/compare/3.6.0...3.7.0